Restructure some of the files.
