+++ src/geometryCSG.js
@@ -1,6 +1,6 @@
 const THREE = require('three');
 const ThreeBSP = require('./three-js-csg')(THREE);
-const Geometry = require('./geometry').Geometry;
+const Geometry = require('./primitives/geometry').Geometry;
 const work = require('webworkify-webpack');
 const Promise = require('promise-polyfill').default;
 //const work = undefined;
