+++ src/primitives/label.js
